Started with scripting.

+++ src/Squidex.Infrastructure/Tasks/PartitionedActionBlock.cs
@@ -23,22 +23,22 @@ namespace Squidex.Infrastructure.Tasks
             get { return Task.WhenAll(workers.Select(x => x.Completion)); }
         }
 
-        public PartitionedActionBlock(Action<TInput> action, Func<TInput, int> partitioner)
+        public PartitionedActionBlock(Action<TInput> action, Func<TInput, long> partitioner)
             : this (action?.ToAsync(), partitioner, new ExecutionDataflowBlockOptions())
         {
         }
 
-        public PartitionedActionBlock(Func<TInput, Task> action, Func<TInput, int> partitioner)
+        public PartitionedActionBlock(Func<TInput, Task> action, Func<TInput, long> partitioner)
             : this(action, partitioner, new ExecutionDataflowBlockOptions())
         {
         }
 
-        public PartitionedActionBlock(Action<TInput> action, Func<TInput, int> partitioner, ExecutionDataflowBlockOptions dataflowBlockOptions)
+        public PartitionedActionBlock(Action<TInput> action, Func<TInput, long> partitioner, ExecutionDataflowBlockOptions dataflowBlockOptions)
             : this(action?.ToAsync(), partitioner, dataflowBlockOptions)
         {
         }
 
-        public PartitionedActionBlock(Func<TInput, Task> action, Func<TInput, int> partitioner, ExecutionDataflowBlockOptions dataflowBlockOptions)
+        public PartitionedActionBlock(Func<TInput, Task> action, Func<TInput, long> partitioner, ExecutionDataflowBlockOptions dataflowBlockOptions)
         {
             Guard.NotNull(action, nameof(action));
             Guard.NotNull(partitioner, nameof(partitioner));
